improve mocha results processing
