+++ test/setup/setup.js
@@ -20,26 +20,6 @@ module.exports = function(mochaRunner) {
 
   var $fixtures;
 
-  function mochaResults(runner) {
-    var failedTests = [];
-
-    runner.on('end', function() {
-      global.mochaResults = runner.stats;
-      global.mochaResults.reports = failedTests;
-    });
-
-    runner.on('fail', function(test, err) {
-      failedTests.push({
-        title: test.title,
-        fullTitle: test.fullTitle(),
-        error: {
-          message: err.message,
-          stack: err.stack
-        }
-      });
-    });
-  }
-
   function setFixtures() {
     _.each(arguments, function(content) {
       $fixtures.append(content);
@@ -62,8 +42,6 @@ module.exports = function(mochaRunner) {
     this.checkProperties = checkProperties;
     this.setFixtures   = setFixtures;
     this.clearFixtures = clearFixtures;
-
-    mochaResults(mochaRunner);
   });
 
   beforeEach(function() {
